Promena iz double u decimal i dodavanje not null ogranicenja u mapiranju voznje i dogadjaja

+++ DrugiDeo/FluentNHibernate/Entiteti/Voznja.cs
@@ -11,14 +11,14 @@ namespace FluentNHibernateTemplate.Entiteti
         public virtual int Id { get; set; }
         public virtual DateTime VremePocetka { get; set; }
         public virtual DateTime VremeZavrsetka { get; set; }
-        public virtual double PredjenaKilometraza { get; set; }
+        public virtual decimal PredjenaKilometraza { get; set; }
         public virtual int TrajanjeMinuti { get; set; }
-        public virtual double PocetniNivo { get; set; }
-        public virtual double KrajnjiNivo { get; set; }
+        public virtual decimal PocetniNivo { get; set; }
+        public virtual decimal KrajnjiNivo { get; set; }
         public virtual string PocetnaLokacija { get; set; }
         public virtual string KrajnjaLokacija { get; set; }
-        public virtual double Cena { get; set; }
-        public virtual double Naknade { get; set; }
+        public virtual decimal Cena { get; set; }
+        public virtual decimal Naknade { get; set; }
         public virtual Rezervacija Rezervacija { get; set; }
 
         public virtual IList<DogadjajUVoznji> Dogadjaji { get; set; }
